+++ __init__.py
@@ -1,22 +1,22 @@
 from typing import Any
 
-from .advanced.blender import AdvancedBlender
-from .advanced.box import AdvancedBox
-from .advanced.download_image import DownloadImage
-from .advanced.download_mask import DownloadMask
-from .advanced.eraser import AdvancedEraser
-from .advanced.recolor import AdvancedRecolor
-from .advanced.segment import AdvancedSegment
-from .advanced.shadow import AdvancedShadow
-from .advanced.upload_image import UploadImage
-from .advanced.upload_mask import UploadMask
-from .skills.blender import Blender
-from .skills.box import Box
-from .skills.eraser import Eraser
-from .skills.name import InferMainSubject
-from .skills.recolor import Recolor
-from .skills.segment import Segment
-from .skills.shadow import Shadow
+from .high_level.blender import Blender
+from .high_level.box import Box
+from .high_level.eraser import Eraser
+from .high_level.name import InferMainSubject
+from .high_level.recolor import Recolor
+from .high_level.segment import Segment
+from .high_level.shadow import Shadow
+from .low_level.blender import AdvancedBlender
+from .low_level.box import AdvancedBox
+from .low_level.download_image import DownloadImage
+from .low_level.download_mask import DownloadMask
+from .low_level.eraser import AdvancedEraser
+from .low_level.recolor import AdvancedRecolor
+from .low_level.segment import AdvancedSegment
+from .low_level.shadow import AdvancedShadow
+from .low_level.upload_image import UploadImage
+from .low_level.upload_mask import UploadMask
 from .utils.bbox import CreateBoundingBox, DrawBoundingBox, ImageCropBoundingBox, MaskCropBoundingBox
 from .utils.image import ApplyTransparencyMask
