+++ high_level/eraser.py
