+++ high_level/shadow.py
